Fix warnings.

+++ modules/tracker/mbt/test/testTukeyEstimator.cpp
@@ -47,7 +47,7 @@
 #include <visp3/mbt/vpMbtTukeyEstimator.h>
 
 
-int main(int /*argc*/, const char **/*argv*/) {
+int main(int /*argc*/, const char ** /*argv*/) {
   size_t nb_elements = 1000;
   int nb_iterations = 100;
   double stdev = 0.5, mean = 0.0, noise_threshold = 1e-3;
@@ -58,7 +58,7 @@ int main(int /*argc*/, const char **/*argv*/) {
   vpColVector residues_col((unsigned int) nb_elements);
   vpColVector weights_col((unsigned int) nb_elements, 1.0), weights_col_save;
   for (size_t i = 0; i < nb_elements; i++) {
-    residues_col[i] = noise();
+    residues_col[(unsigned int) i] = noise();
   }
 
   vpRobust robust((unsigned int) nb_elements);
@@ -98,7 +98,7 @@ int main(int /*argc*/, const char **/*argv*/) {
 
   //Generate again for weights != 1
   for (size_t i = 0; i < nb_elements; i++) {
-    residues_col[i] = noise();
+    residues_col[(unsigned int) i] = noise();
   }
   weights_col_save = weights_col;
   t_robust = vpTime::measureTimeMs();
@@ -118,7 +118,7 @@ int main(int /*argc*/, const char **/*argv*/) {
 
     double t = vpTime::measureTimeMs();
     for (int i = 0; i < nb_iterations; i++) {
-      tukey_estimator.MEstimator(residues, weights, noise_threshold);
+      tukey_estimator.MEstimator(residues, weights, (float) noise_threshold);
     }
     t = vpTime::measureTimeMs() - t;
 
@@ -136,7 +136,7 @@ int main(int /*argc*/, const char **/*argv*/) {
 
   //Generate again for weights != 1 and vpColVector type
   for (size_t i = 0; i < nb_elements; i++) {
-    residues_col[i] = noise();
+    residues_col[(unsigned int) i] = noise();
   }
   weights_col_save = weights_col;
   t_robust = vpTime::measureTimeMs();
